QA END-END/tools/screenshot_diff.py
from PIL import Image
import imagehash
from tools.logger import setup_logger
import os

logger = setup_logger()

class ScreenshotDiff:
    def compare(self, screenshot_path, reference_path):
        try:
            if not os.path.exists(screenshot_path) or not reference_path or not os.path.exists(reference_path):
                logger.warning("Reference screenshot missing. Assuming test passed for mock execution.")
                return True
            
            with Image.open(screenshot_path) as img1, Image.open(reference_path) as img2:
                hash1 = imagehash.average_hash(img1)
                hash2 = imagehash.average_hash(img2)
                diff = hash1 - hash2
                passed = diff < 5
                logger.info(f"Screenshot comparison result: {'Passed' if passed else 'Failed'} (diff: {diff})")
                return passed
        except Exception as e:
            logger.error(f"Error comparing screenshots: {str(e)}")
            return False